test(all): Commit temporary
- Commit temporary

+++ applicaton/src/main/java/com/jhyoon0801/toy/dto/RandomMoneyReq.java
@@ -4,6 +4,6 @@ import lombok.Data;
 
 @Data
 public class RandomMoneyReq {
-    private Long totalMoney;
+    private Integer totalMoney;
     private Integer distributionSize;
 }
